Issue #722 Refactor EhcacheManager to separate persistence
Refactors EhcacheManager to split CachePersistenceManager methods out.
This creates a BaseCacheManager abstract class holding the non-persistent
methods leaving EhcacheManager as a subclass maintaining the current
functionality.  This is a positioning move for supporting clustering.

* Adds requiresSerialization method to ResourceType
* Splits EhcacheManager.newStore into BaseCacheManager (non-DISK)
* Generalizes CacheManagerBuilder to
** return builder-designated CacheManager type (subclass)
** return CacheManagerConfiguration-driven CacheManagerBuilder subclass
** adds newSpecializedCacheManager method to obtain type-specific
   CacheManager from configuration
