+++ core/src/main/java/org/ehcache/core/BaseCacheManager.java
@@ -1,0 +1,634 @@
+/*
+ * Copyright Terracotta, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.ehcache.core;
+
+import org.ehcache.Cache;
+import org.ehcache.CacheManager;
+import org.ehcache.Status;
+import org.ehcache.config.CacheConfiguration;
+import org.ehcache.config.Configuration;
+import org.ehcache.config.ResourceType;
+import org.ehcache.config.RuntimeConfiguration;
+import org.ehcache.core.config.BaseCacheConfiguration;
+import org.ehcache.core.config.DefaultConfiguration;
+import org.ehcache.core.config.events.StoreEventSourceConfiguration;
+import org.ehcache.core.config.store.StoreConfigurationImpl;
+import org.ehcache.core.events.CacheEventDispatcher;
+import org.ehcache.core.events.CacheEventDispatcherFactory;
+import org.ehcache.core.events.CacheManagerListener;
+import org.ehcache.core.spi.LifeCycledAdapter;
+import org.ehcache.core.spi.ServiceLocator;
+import org.ehcache.core.spi.cache.Store;
+import org.ehcache.core.spi.service.CacheManagerProviderService;
+import org.ehcache.core.util.ClassLoading;
+import org.ehcache.event.CacheEventListener;
+import org.ehcache.event.CacheEventListenerConfiguration;
+import org.ehcache.event.CacheEventListenerProvider;
+import org.ehcache.spi.LifeCycled;
+import org.ehcache.spi.cache.ObservableCacheManager;
+import org.ehcache.spi.loaderwriter.CacheLoaderWriter;
+import org.ehcache.spi.loaderwriter.CacheLoaderWriterProvider;
+import org.ehcache.spi.loaderwriter.WriteBehindConfiguration;
+import org.ehcache.spi.loaderwriter.WriteBehindProvider;
+import org.ehcache.spi.serialization.SerializationProvider;
+import org.ehcache.spi.serialization.Serializer;
+import org.ehcache.spi.serialization.UnsupportedTypeException;
+import org.ehcache.spi.service.Service;
+import org.ehcache.spi.service.ServiceConfiguration;
+import org.ehcache.spi.service.ServiceCreationConfiguration;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.ArrayDeque;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Deque;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.CopyOnWriteArrayList;
+
+/**
+ * Support the common core of {@code CacheManager} functionality.
+ *
+ * @author Clifford W. Johnson
+ */
+public abstract class BaseCacheManager implements CacheManager, ObservableCacheManager {
+
+  private final DefaultConfiguration configuration;
+  private final ClassLoader cacheManagerClassLoader;
+  private final boolean useLoaderInAtomics;
+
+  private final Logger logger = LoggerFactory.getLogger(this.getClass());
+  private final ConcurrentMap<String, CacheHolder> caches = new ConcurrentHashMap<String, CacheHolder>();
+  private final CopyOnWriteArrayList<CacheManagerListener> listeners = new CopyOnWriteArrayList<CacheManagerListener>();
+
+  protected final StatusTransitioner statusTransitioner = new StatusTransitioner(this.logger);
+  protected final String simpleName;
+  protected final ServiceLocator serviceLocator;
+
+  protected BaseCacheManager(Configuration config, Collection<Service> services, boolean useLoaderInAtomics) {
+    final String simpleName = this.getClass().getSimpleName();
+    this.simpleName = (simpleName.isEmpty() ? this.getClass().getName() : simpleName);
+    this.configuration = new DefaultConfiguration(config);
+    this.cacheManagerClassLoader = config.getClassLoader() != null ? config.getClassLoader() : ClassLoading.getDefaultClassLoader();
+    this.serviceLocator = new ServiceLocator(services.toArray(new Service[services.size()]));
+    this.useLoaderInAtomics = useLoaderInAtomics;
+    validateServicesConfigs();
+  }
+
+  private void validateServicesConfigs() {
+    HashSet<Class> classes = new HashSet<Class>();
+    for (ServiceCreationConfiguration<?> service : configuration.getServiceCreationConfigurations()) {
+      if (!classes.add(service.getServiceType())) {
+        throw new IllegalStateException("Duplicate creation configuration for service " + service.getServiceType());
+      }
+    }
+  }
+
+  /**
+   * Gets the class-identified {@code Logger} instance.
+   *
+   * @return the {@code Logger} to use
+   */
+  protected Logger getLogger() {
+    return this.logger;
+  }
+
+  /**
+   * Gets the {@code Class} definition holding the {@link org.ehcache.spi.service.ServiceDependencies ServiceDependencies}
+   * annotations for this {@code CacheManager} implementation.
+   *
+   * @return the {@code Class} definition from which the service dependencies are identified
+   */
+  protected abstract Class<?> getServiceDependencyMarkerClass();
+
+  @Override
+  public <K, V> Cache<K, V> getCache(String alias, Class<K> keyType, Class<V> valueType) {
+    statusTransitioner.checkAvailable();
+    final CacheHolder cacheHolder = caches.get(alias);
+    if(cacheHolder == null) {
+      return null;
+    } else {
+      try {
+        return cacheHolder.retrieve(keyType, valueType);
+      } catch (IllegalArgumentException e) {
+        throw new IllegalArgumentException("Cache '" + alias + "' type is <" + cacheHolder.keyType.getName() + ", "
+                                           + cacheHolder.valueType.getName() + ">, but you retrieved it with <"
+                                           + keyType.getName() + ", " + valueType.getName() +">");
+      }
+    }
+  }
+
+  @Override
+  public void removeCache(final String alias) {
+    removeCache(alias, true);
+  }
+
+  /**
+   * Closes and removes a cache, by alias, from this cache manager.
+   *
+   * @param alias the alias of the cache to remove
+   * @param removeFromConfig if {@code true}, the cache configuration is altered to remove the cache
+   */
+  private void removeCache(final String alias, final boolean removeFromConfig) {
+    statusTransitioner.checkAvailable();
+    final CacheHolder cacheHolder = caches.remove(alias);
+    if(cacheHolder != null) {
+      final Ehcache<?, ?> ehcache = cacheHolder.retrieve(cacheHolder.keyType, cacheHolder.valueType);
+      if(!statusTransitioner.isTransitioning()) {
+        for (CacheManagerListener listener : listeners) {
+          listener.cacheRemoved(alias, ehcache);
+        }
+      }
+      closeCache(alias, ehcache);
+      if (removeFromConfig) {
+        configuration.removeCacheConfiguration(alias);
+      }
+      this.getLogger().info("Cache '{}' is removed from {}.", alias, simpleName);
+    }
+  }
+
+  private void closeCache(final String alias, final Ehcache<?, ?> ehcache) {
+    ehcache.close();
+    closeEhcache(alias, ehcache);
+    this.getLogger().info("Cache '{}' is closed from {}.", alias, simpleName);
+  }
+
+  /**
+   * Perform cache closure actions specific to a cache manager implementation.
+   * This method is called <i>after</i> the {@code Ehcache} instance is closed.
+   *
+   * @param alias the cache alias
+   * @param ehcache the {@code Ehcache} instance for the cache to close
+   */
+  protected abstract void closeEhcache(final String alias, final Ehcache<?, ?> ehcache);
+
+  @Override
+  public <K, V> Cache<K, V> createCache(final String alias, CacheConfiguration<K, V> config) throws IllegalArgumentException {
+    return createCache(alias, config, true);
+  }
+
+  private <K, V> Cache<K, V> createCache(final String alias, CacheConfiguration<K, V> originalConfig, boolean addToConfig) throws IllegalArgumentException {
+    statusTransitioner.checkAvailable();
+
+    this.getLogger().info("Cache '{}' is getting created in {}.", alias, simpleName);
+
+    CacheConfiguration<K, V> config = adjustConfigurationWithCacheManagerDefaults(originalConfig);
+    Class<K> keyType = config.getKeyType();
+    Class<V> valueType = config.getValueType();
+
+    final CacheHolder value = new CacheHolder(keyType, valueType, null);
+    if (caches.putIfAbsent(alias, value) != null) {
+      throw new IllegalArgumentException("Cache '" + alias +"' already exists");
+    }
+
+    Ehcache<K, V> cache = null;
+
+    RuntimeException failure = null;
+    try {
+      cache = createNewEhcache(alias, config, keyType, valueType);
+      cache.init();
+      if (addToConfig) {
+        configuration.addCacheConfiguration(alias, cache.getRuntimeConfiguration());
+      } else {
+        configuration.replaceCacheConfiguration(alias, originalConfig, cache.getRuntimeConfiguration());
+      }
+    } catch (RuntimeException e) {
+      failure = e;
+    }
+
+    if(failure == null) {
+      try {
+        if(!statusTransitioner.isTransitioning()) {
+          for (CacheManagerListener listener : listeners) {
+            listener.cacheAdded(alias, cache);
+          }
+        }
+      } finally {
+        value.setCache(cache);
+      }
+    } else {
+      caches.remove(alias);
+      value.setCache(null);
+      throw new IllegalStateException("Cache '"+alias+"' creation in " + simpleName +
+          " failed.", failure);
+    }
+    this.getLogger().info("Cache '{}' created in {}.", alias, simpleName);
+    return cache;
+  }
+
+  <K, V> Ehcache<K, V> createNewEhcache(final String alias, final CacheConfiguration<K, V> config,
+                                        final Class<K> keyType, final Class<V> valueType) {
+    Collection<ServiceConfiguration<?>> adjustedServiceConfigs = new ArrayList<ServiceConfiguration<?>>(config.getServiceConfigurations());
+
+    List<ServiceConfiguration> unknownServiceConfigs = new ArrayList<ServiceConfiguration>();
+    for (ServiceConfiguration serviceConfig : adjustedServiceConfigs) {
+      if (!serviceLocator.knowsServiceFor(serviceConfig)) {
+        unknownServiceConfigs.add(serviceConfig);
+      }
+    }
+    if (!unknownServiceConfigs.isEmpty()) {
+      throw new IllegalStateException("Cannot find service(s) that can handle following configuration(s) : " + unknownServiceConfigs);
+    }
+
+    List<LifeCycled> lifeCycledList = new ArrayList<LifeCycled>();
+
+    final Store<K, V> store = getStore(alias, config, keyType, valueType, adjustedServiceConfigs, lifeCycledList);
+
+    final CacheLoaderWriterProvider cacheLoaderWriterProvider = serviceLocator.getService(CacheLoaderWriterProvider.class);
+    final CacheLoaderWriter<? super K, V> decorator ;
+    if(cacheLoaderWriterProvider != null) {
+      final CacheLoaderWriter<? super K, V> loaderWriter;
+      loaderWriter = cacheLoaderWriterProvider.createCacheLoaderWriter(alias, config);
+      WriteBehindConfiguration writeBehindConfiguration =
+          ServiceLocator.findSingletonAmongst(WriteBehindConfiguration.class, config.getServiceConfigurations().toArray());
+      if(writeBehindConfiguration == null) {
+        decorator = loaderWriter;
+      } else {
+        final WriteBehindProvider factory = serviceLocator.getService(WriteBehindProvider.class);
+        decorator = factory.createWriteBehindLoaderWriter(loaderWriter, writeBehindConfiguration);
+        if(decorator != null) {
+          lifeCycledList.add(new LifeCycledAdapter() {
+            @Override
+            public void close() {
+              factory.releaseWriteBehindLoaderWriter(decorator);
+            }
+          });
+        }
+      }
+
+      if (loaderWriter != null) {
+        lifeCycledList.add(new LifeCycledAdapter() {
+          @Override
+          public void close() throws Exception {
+            cacheLoaderWriterProvider.releaseCacheLoaderWriter(loaderWriter);
+          }
+        });
+      }
+    } else {
+      decorator = null;
+    }
+
+    final CacheEventDispatcherFactory cenlProvider = serviceLocator.getService(CacheEventDispatcherFactory.class);
+    final CacheEventDispatcher<K, V> evtService =
+        cenlProvider.createCacheEventDispatcher(store, adjustedServiceConfigs.toArray(new ServiceConfiguration[adjustedServiceConfigs.size()]));
+    lifeCycledList.add(new LifeCycledAdapter() {
+      @Override
+      public void close() {
+        cenlProvider.releaseCacheEventDispatcher(evtService);
+      }
+    });
+    evtService.setStoreEventSource(store.getStoreEventSource());
+
+    final Ehcache<K, V> ehCache = new Ehcache<K, V>(config, store, decorator, evtService,
+        useLoaderInAtomics, LoggerFactory.getLogger(Ehcache.class + "-" + alias));
+
+    final CacheEventListenerProvider evntLsnrFactory = serviceLocator.getService(CacheEventListenerProvider.class);
+    if (evntLsnrFactory != null) {
+      Collection<CacheEventListenerConfiguration> evtLsnrConfigs =
+          ServiceLocator.findAmongst(CacheEventListenerConfiguration.class, config.getServiceConfigurations());
+      for (CacheEventListenerConfiguration lsnrConfig: evtLsnrConfigs) {
+        final CacheEventListener<K, V> lsnr = evntLsnrFactory.createEventListener(alias, lsnrConfig);
+        if (lsnr != null) {
+          ehCache.getRuntimeConfiguration().registerCacheEventListener(lsnr, lsnrConfig.orderingMode(), lsnrConfig.firingMode(),
+              lsnrConfig.fireOn());
+          lifeCycledList.add(new LifeCycled() {
+            @Override
+            public void init() throws Exception {
+              // no-op for now
+            }
+
+            @Override
+            public void close() throws Exception {
+              evntLsnrFactory.releaseEventListener(lsnr);
+            }
+          });
+        }
+      }
+      evtService.setListenerSource(ehCache);
+    }
+
+    for (LifeCycled lifeCycled : lifeCycledList) {
+      ehCache.addHook(lifeCycled);
+    }
+
+    return ehCache;
+  }
+
+  /**
+   * Instantiates a {@code Store} used for the cache data.
+   *
+   * @param alias the alias assigned to the cache
+   * @param config the configuration used for the cache
+   * @param keyType the cache key type
+   * @param valueType the cache value type
+   * @param serviceConfigs the {@code List} of {@code ServiceConfiguration} instances available to the cache;
+   *                       this list may be augmented by the implementation of this method
+   * @param lifeCycledList the {@code List} of {@code LifeCycled} instances used to manage components of the
+   *                       cache; this list may be augmented by the implementation of this method
+   * @param <K> the cache key type
+   * @param <V> the cache value type
+   *
+   * @return the {@code Store} instance used to create the cache
+   */
+  protected <K, V> Store<K,V> getStore(final String alias, final CacheConfiguration<K, V> config,
+                                       final Class<K> keyType, final Class<V> valueType,
+                                       final Collection<ServiceConfiguration<?>> serviceConfigs,
+                                       final List<LifeCycled> lifeCycledList) {
+
+    final Store.Provider storeProvider = serviceLocator.getService(Store.Provider.class);
+    Serializer<K> keySerializer = null;
+    Serializer<V> valueSerializer = null;
+    final SerializationProvider serialization = serviceLocator.getService(SerializationProvider.class);
+    Set<ResourceType> resources = config.getResourcePools().getResourceTypeSet();
+    ServiceConfiguration<?>[] serviceConfigArray = serviceConfigs.toArray(new ServiceConfiguration[serviceConfigs.size()]);
+    if (serialization != null) {
+      try {
+        final Serializer<K> keySer = serialization.createKeySerializer(keyType, config.getClassLoader(), serviceConfigArray);
+        lifeCycledList.add(new LifeCycledAdapter() {
+          @Override
+          public void close() throws Exception {
+            serialization.releaseSerializer(keySer);
+          }
+        });
+        keySerializer = keySer;
+      } catch (UnsupportedTypeException e) {
+        for (ResourceType resource : resources) {
+          if (resource.requiresSerialization()) {
+            throw new RuntimeException(e);
+          }
+        }
+        this.getLogger().debug("Could not create serializers for " + alias, e);
+      }
+      try {
+        final Serializer<V> valueSer = serialization.createValueSerializer(valueType, config.getClassLoader(), serviceConfigArray);
+        lifeCycledList.add(new LifeCycledAdapter() {
+          @Override
+          public void close() throws Exception {
+            serialization.releaseSerializer(valueSer);
+          }
+        });
+        valueSerializer = valueSer;
+      } catch (UnsupportedTypeException e) {
+        for (ResourceType resource : resources) {
+          if (resource.requiresSerialization()) {
+            throw new RuntimeException(e);
+          }
+        }
+        this.getLogger().debug("Could not create serializers for " + alias, e);
+      }
+    }
+
+    int eventParallelism;
+    StoreEventSourceConfiguration eventSourceConfiguration = ServiceLocator.findSingletonAmongst(StoreEventSourceConfiguration.class, config
+        .getServiceConfigurations()
+        .toArray());
+    if (eventSourceConfiguration != null) {
+      eventParallelism = eventSourceConfiguration.getOrderedEventParallelism();
+    } else {
+      eventParallelism = StoreEventSourceConfiguration.DEFAULT_EVENT_PARALLELISM;
+    }
+
+    Store.Configuration<K, V> storeConfiguration = new StoreConfigurationImpl<K, V>(config, eventParallelism, keySerializer, valueSerializer);
+    final Store<K, V> store = storeProvider.createStore(storeConfiguration, serviceConfigArray);
+
+    lifeCycledList.add(new LifeCycled() {
+      @Override
+      public void init() throws Exception {
+        storeProvider.initStore(store);
+      }
+
+      @Override
+      public void close() {
+        storeProvider.releaseStore(store);
+      }
+    });
+
+    return store;
+  }
+
+  /**
+   *  adjusts the config to reflect new classloader & serialization provider
+   */
+  private <K, V> CacheConfiguration<K, V> adjustConfigurationWithCacheManagerDefaults(CacheConfiguration<K, V> config) {
+    ClassLoader cacheClassLoader = config.getClassLoader();
+    if (cacheClassLoader == null) {
+      cacheClassLoader = cacheManagerClassLoader;
+    }
+    if (cacheClassLoader != config.getClassLoader() ) {
+      config = new BaseCacheConfiguration<K, V>(config.getKeyType(), config.getValueType(),
+          config.getEvictionVeto(), cacheClassLoader, config.getExpiry(),
+          config.getResourcePools(), config.getServiceConfigurations().toArray(
+          new ServiceConfiguration<?>[config.getServiceConfigurations().size()]));
+    }
+    return config;
+  }
+
+  @Override
+  public void registerListener(CacheManagerListener listener) {
+    if(!listeners.contains(listener)) {
+      listeners.add(listener);
+      statusTransitioner.registerListener(listener);
+    }
+  }
+
+  @Override
+  public void deregisterListener(CacheManagerListener listener) {
+    if(listeners.remove(listener)) {
+      statusTransitioner.deregisterListener(listener);
+    }
+  }
+
+  /**
+   * {@inheritDoc}
+   */
+  @Override
+  public void init() {
+    final StatusTransitioner.Transition st = statusTransitioner.init();
+
+    if (serviceLocator.getService(CacheManagerProviderService.class) == null) {
+      this.serviceLocator.addService(new DefaultCacheManagerProviderService(this));
+    }
+
+    try {
+      for (ServiceCreationConfiguration<? extends Service> serviceConfig : configuration.getServiceCreationConfigurations()) {
+        Service service = serviceLocator.getOrCreateServiceFor(serviceConfig);
+        if (service == null) {
+          throw new IllegalArgumentException("Couldn't resolve Service " + serviceConfig.getServiceType().getName());
+        }
+      }
+      serviceLocator.loadDependenciesOf(this.getServiceDependencyMarkerClass());
+      try {
+        serviceLocator.startAllServices();
+      } catch (Exception e) {
+        throw st.failed(e);
+      }
+
+      Deque<String> initiatedCaches = new ArrayDeque<String>();
+      try {
+        for (Map.Entry<String, CacheConfiguration<?, ?>> cacheConfigurationEntry : configuration.getCacheConfigurations()
+            .entrySet()) {
+          final String alias = cacheConfigurationEntry.getKey();
+          createCache(alias, cacheConfigurationEntry.getValue(), false);
+          initiatedCaches.push(alias);
+        }
+      } catch (RuntimeException e) {
+        while (!initiatedCaches.isEmpty()) {
+          String toBeClosed = initiatedCaches.pop();
+          try {
+            removeCache(toBeClosed, false);
+          } catch (Exception exceptionClosingCache) {
+              this.getLogger().error("Cache '{}' could not be removed due to ", toBeClosed, exceptionClosingCache);
+          }
+        }
+        try {
+          serviceLocator.stopAllServices();
+        } catch (Exception exceptionStoppingServices) {
+          this.getLogger().error("Stopping services failed due to ", exceptionStoppingServices);
+        }
+        throw e;
+      }
+    } catch (Exception e) {
+      throw st.failed(e);
+    }
+    st.succeeded();
+  }
+
+  @Override
+  public Status getStatus() {
+    return statusTransitioner.currentStatus();
+  }
+
+  @Override
+  public void close() {
+    final StatusTransitioner.Transition st = statusTransitioner.close();
+
+    Exception firstException = null;
+    try {
+      for (String alias : caches.keySet()) {
+        try {
+          removeCache(alias, false);
+        } catch (Exception e) {
+          if(firstException == null) {
+            firstException = e;
+          } else {
+            this.getLogger().error("Cache '{}' could not be removed due to ", alias, e);
+          }
+        }
+      }
+
+      serviceLocator.stopAllServices();
+    } catch (Exception e) {
+      if(firstException == null) {
+        firstException = e;
+      }
+    }
+    if(firstException != null) {
+      throw st.failed(firstException);
+    }
+    st.succeeded();
+  }
+
+  @Override
+  public RuntimeConfiguration getRuntimeConfiguration() {
+    return configuration;
+  }
+
+  /**
+   * Removes and closes a cache without performing {@link CacheManagerListener#cacheRemoved(String, Cache)}
+   * notifications.
+   *
+   * @param alias the alias of the cache to remove
+   */
+  protected void removeAndCloseWithoutNotice(final String alias) {
+    final CacheHolder cacheHolder = caches.remove(alias);
+    if(cacheHolder != null) {
+      final Ehcache<?, ?> ehcache = cacheHolder.retrieve(cacheHolder.keyType, cacheHolder.valueType);
+      if(ehcache.getStatus() == Status.AVAILABLE) {
+        ehcache.close();
+      }
+    }
+  }
+
+  // for tests at the moment
+  ClassLoader getClassLoader() {
+    return cacheManagerClassLoader;
+  }
+
+  void create() {
+    statusTransitioner.checkMaintenance();
+  }
+
+  void destroy() {
+    statusTransitioner.checkMaintenance();
+  }
+
+
+  private static final class CacheHolder {
+    private final Class<?> keyType;
+    private final Class<?> valueType;
+    private volatile Ehcache<?, ?> cache;
+    private volatile boolean isValueSet = false;
+
+    CacheHolder(Class<?> keyType, Class<?> valueType, Ehcache<?, ?> cache) {
+      this.keyType = keyType;
+      this.valueType = valueType;
+      this.cache = cache;
+    }
+
+    <K, V> Ehcache<K, V> retrieve(Class<K> refKeyType, Class<V> refValueType) {
+      if (!isValueSet) {
+        synchronized (this) {
+          boolean interrupted = false;
+          try {
+            while(!isValueSet) {
+              try {
+                wait();
+              } catch (InterruptedException e) {
+                interrupted = true;
+              }
+            }
+          } finally {
+            if(interrupted) {
+              Thread.currentThread().interrupt();
+            }
+          }
+        }
+      }
+      if (keyType == refKeyType && valueType == refValueType) {
+        return cast(cache);
+      } else {
+        throw new IllegalArgumentException();
+      }
+    }
+
+    @SuppressWarnings("unchecked")
+    private static <K, V> Ehcache<K, V> cast(Ehcache<?, ?> cache) {
+      return (Ehcache<K, V>)cache;
+    }
+
+    public synchronized void setCache(final Ehcache<?, ?> cache) {
+      this.cache = cache;
+      this.isValueSet = true;
+      notifyAll();
+    }
+  }
+}
