+++ impl/src/main/java/org/ehcache/config/builders/CacheManagerBuilder.java
@@ -17,6 +17,7 @@
 package org.ehcache.config.builders;
 
 import org.ehcache.CacheManager;
+import org.ehcache.PersistentCacheManager;
 import org.ehcache.config.CacheConfiguration;
 import org.ehcache.config.Configuration;
 import org.ehcache.config.units.MemoryUnit;
@@ -38,6 +39,7 @@ import org.ehcache.spi.service.ServiceCreationConfiguration;
 import java.io.File;
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
 import static java.util.Collections.emptySet;
@@ -68,37 +70,46 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
     return build(false);
   }
 
-  private CacheManagerBuilder() {
+  protected CacheManagerBuilder() {
     this.configBuilder = newConfigurationBuilder();
     this.services = emptySet();
   }
 
-  private CacheManagerBuilder(CacheManagerBuilder<T> builder, Set<Service> services) {
+  protected CacheManagerBuilder(CacheManagerBuilder<T> builder, Set<Service> services) {
     this.configBuilder = builder.configBuilder;
     this.services = unmodifiableSet(services);
   }
 
-  private CacheManagerBuilder(CacheManagerBuilder<T> builder, ConfigurationBuilder configBuilder) {
+  protected CacheManagerBuilder(CacheManagerBuilder<T> builder, ConfigurationBuilder configBuilder) {
     this.configBuilder = configBuilder;
     this.services = builder.services;
   }
 
-  public static CacheManager newCacheManager(final Configuration configuration) {
-    return new EhcacheManager(configuration);
+  protected CacheManagerBuilder(final CacheManagerBuilder<?> builder) {
+    this.configBuilder = builder.configBuilder;
+    this.services = builder.services;
   }
 
-  T newCacheManager(Collection<Service> services, final Configuration configuration) {
+  protected T newCacheManager(Collection<Service> services, final Configuration configuration) {
     final EhcacheManager ehcacheManager = new EhcacheManager(configuration, services);
     return cast(ehcacheManager);
   }
 
+  protected CacheManagerBuilder<T> newBuilder(final Set<Service> services) {
+    return new CacheManagerBuilder<T>(this, services);
+  }
+
+  protected CacheManagerBuilder<T> newBuilder(final ConfigurationBuilder builder) {
+    return new CacheManagerBuilder<T>(this, builder);
+  }
+
   @SuppressWarnings("unchecked")
-  T cast(EhcacheManager ehcacheManager) {
+  private T cast(EhcacheManager ehcacheManager) {
     return (T) ehcacheManager;
   }
 
   public <K, V> CacheManagerBuilder<T> withCache(String alias, CacheConfiguration<K, V> configuration) {
-    return new CacheManagerBuilder<T>(this, configBuilder.addCache(alias, configuration));
+    return this.newBuilder(configBuilder.addCache(alias, configuration));
   }
 
   public <K, V> CacheManagerBuilder<T> withCache(String alias, CacheConfigurationBuilder<K, V> configurationBuilder) {
@@ -112,7 +123,7 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
   public CacheManagerBuilder<T> using(Service service) {
     Set<Service> newServices = new HashSet<Service>(services);
     newServices.add(service);
-    return new CacheManagerBuilder<T>(this, newServices);
+    return this.newBuilder(newServices);
   }
 
   public <C> CacheManagerBuilder<T> withCopier(Class<C> clazz, Class<? extends Copier<C>> copier) {
@@ -120,11 +131,11 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
     if (service == null) {
       service = new DefaultCopyProviderConfiguration();
       service.addCopierFor(clazz, copier);
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(service));
+      return this.newBuilder(configBuilder.addService(service));
     } else {
       DefaultCopyProviderConfiguration newConfig = new DefaultCopyProviderConfiguration(service);
       newConfig.addCopierFor(clazz, copier, true);
-      return new CacheManagerBuilder<T>(this, configBuilder.removeService(service).addService(newConfig));
+      return this.newBuilder(configBuilder.removeService(service).addService(newConfig));
     }
   }
 
@@ -133,11 +144,11 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
     if (service == null) {
       service = new DefaultSerializationProviderConfiguration();
       service.addSerializerFor(clazz, serializer);
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(service));
+      return this.newBuilder(configBuilder.addService(service));
     } else {
       DefaultSerializationProviderConfiguration newConfig = new DefaultSerializationProviderConfiguration(service);
       newConfig.addSerializerFor(clazz, serializer, true);
-      return new CacheManagerBuilder<T>(this, configBuilder.removeService(service).addService(newConfig));
+      return this.newBuilder(configBuilder.removeService(service).addService(newConfig));
     }
   }
 
@@ -154,55 +165,62 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
   public CacheManagerBuilder<T> withDefaultSizeOfMaxObjectSize(long size, MemoryUnit unit) {
     SizeOfEngineProviderConfiguration configuration = configBuilder.findServiceByClass(DefaultSizeOfEngineProviderConfiguration.class);
     if (configuration == null) {
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(new DefaultSizeOfEngineProviderConfiguration(size, unit, DEFAULT_OBJECT_GRAPH_SIZE)));
+      return this.newBuilder(configBuilder.addService(new DefaultSizeOfEngineProviderConfiguration(size, unit, DEFAULT_OBJECT_GRAPH_SIZE)));
     } else {
       ConfigurationBuilder builder = configBuilder.removeService(configuration);
-      return new CacheManagerBuilder<T>(this, builder.addService(new DefaultSizeOfEngineProviderConfiguration(size, unit, configuration.getMaxObjectGraphSize())));
+      return this.newBuilder(builder.addService(new DefaultSizeOfEngineProviderConfiguration(size, unit, configuration.getMaxObjectGraphSize())));
     }
   }
 
   public CacheManagerBuilder<T> withDefaultWriteBehindThreadPool(String threadPoolAlias) {
     WriteBehindProviderConfiguration config = configBuilder.findServiceByClass(WriteBehindProviderConfiguration.class);
     if (config == null) {
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(new WriteBehindProviderConfiguration(threadPoolAlias)));
+      return this.newBuilder(configBuilder.addService(new WriteBehindProviderConfiguration(threadPoolAlias)));
     } else {
       ConfigurationBuilder builder = configBuilder.removeService(config);
-      return new CacheManagerBuilder<T>(this, builder.addService(new WriteBehindProviderConfiguration(threadPoolAlias)));
+      return this.newBuilder(builder.addService(new WriteBehindProviderConfiguration(threadPoolAlias)));
     }
   }
 
   public CacheManagerBuilder<T> withDefaultDiskStoreThreadPool(String threadPoolAlias) {
     OffHeapDiskStoreProviderConfiguration config = configBuilder.findServiceByClass(OffHeapDiskStoreProviderConfiguration.class);
     if (config == null) {
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(new OffHeapDiskStoreProviderConfiguration(threadPoolAlias)));
+      return this.newBuilder(configBuilder.addService(new OffHeapDiskStoreProviderConfiguration(threadPoolAlias)));
     } else {
       ConfigurationBuilder builder = configBuilder.removeService(config);
-      return new CacheManagerBuilder<T>(this, builder.addService(new OffHeapDiskStoreProviderConfiguration(threadPoolAlias)));
+      return this.newBuilder(builder.addService(new OffHeapDiskStoreProviderConfiguration(threadPoolAlias)));
     }
   }
 
   public CacheManagerBuilder<T> withDefaultEventListenersThreadPool(String threadPoolAlias) {
     CacheEventDispatcherFactoryConfiguration config = configBuilder.findServiceByClass(CacheEventDispatcherFactoryConfiguration.class);
     if (config == null) {
-      return new CacheManagerBuilder<T>(this, configBuilder.addService(new CacheEventDispatcherFactoryConfiguration(threadPoolAlias)));
+      return this.newBuilder(configBuilder.addService(new CacheEventDispatcherFactoryConfiguration(threadPoolAlias)));
     } else {
       ConfigurationBuilder builder = configBuilder.removeService(config);
-      return new CacheManagerBuilder<T>(this, builder.addService(new CacheEventDispatcherFactoryConfiguration(threadPoolAlias)));
+      return this.newBuilder(builder.addService(new CacheEventDispatcherFactoryConfiguration(threadPoolAlias)));
     }
   }
 
   public CacheManagerBuilder<T> using(ServiceCreationConfiguration<?> serviceConfiguration) {
-    return new CacheManagerBuilder<T>(this, configBuilder.addService(serviceConfiguration));
+    return this.newBuilder(configBuilder.addService(serviceConfiguration));
   }
 
   public CacheManagerBuilder<T> replacing(ServiceCreationConfiguration<?> overwriteServiceConfiguration) {
     ServiceCreationConfiguration existingConfiguration = configBuilder.findServiceByClass(overwriteServiceConfiguration.getClass());
-    return new CacheManagerBuilder<T>(this, configBuilder.removeService(existingConfiguration)
-        .addService(overwriteServiceConfiguration));
+    return this.newBuilder(configBuilder.removeService(existingConfiguration).addService(overwriteServiceConfiguration));
   }
 
   public CacheManagerBuilder<T> withClassLoader(ClassLoader classLoader) {
-    return new CacheManagerBuilder<T>(this, configBuilder.withClassLoader(classLoader));
+    return this.newBuilder(configBuilder.withClassLoader(classLoader));
+  }
+
+  public static CacheManager newCacheManager(final Configuration configuration) {
+    return newSpecializedCacheManager(CacheManager.class, configuration);
+  }
+
+  public static PersistentCacheManager newPersistentCacheManager(final Configuration config) {
+    return newSpecializedCacheManager(PersistentCacheManager.class, config);
   }
 
   public static CacheManagerBuilder<CacheManager> newCacheManagerBuilder() {
@@ -211,5 +229,53 @@ public class CacheManagerBuilder<T extends CacheManager> implements Builder<T> {
 
   public static PersistenceConfiguration persistence(String location) {
     return new CacheManagerPersistenceConfiguration(new File(location));
+  }
+
+  /**
+   * Creates a {@code CacheManager} of the designated type from the configuration provided.
+   * Configurations passed to this method must have either no more than one
+   * {@link CacheManagerConfiguration} present in the
+   * {@linkplain Configuration#getServiceCreationConfigurations() service creation configuration list}
+   * or the {@code CacheManagerConfiguration} instances must be combinable.  When neither of these
+   * constraints is met, the result of this method is not defined.
+   *
+   * @param specialization the type of the desired {@code CacheManager}
+   * @param config the {@code Configuration} to use in the {@code CacheManager}
+   * @param <T> the type of the desired {@code CacheManager}
+   *
+   * @return a new {@code CacheManager} instance of the type designated by {@code specialization};
+   *      the returned {@code CacheManager} is <b>not</b> initialized
+   *
+   * @throws ClassCastException if the {@link CacheManagerBuilder} associated with the
+   *        {@link CacheManagerConfiguration} instances appearing in the configuration does
+   *        not build a {@code CacheManager} of the {@code specialization} type
+   */
+  public static <T extends CacheManager> T newSpecializedCacheManager(Class<T> specialization, Configuration config)
+      throws ClassCastException {
+    CacheManagerBuilder<?> cmBuilder = newCacheManagerBuilder();
+    /*
+     * Cycle though the service configurations looking for those that also implement CacheManagerConfiguration.
+     * For each CacheManagerConfiguration found, direct the current CacheManagerBuilder (through the 'with' method)
+     * to obtain an instance of the builder for the configuration.  If the configuration is *not* a
+     * CacheManagerConfiguration, simply add it to the set of service configurations.
+     *
+     * Through its 'builder' method, a CacheManagerConfiguration has the capability of designating the
+     * CacheManagerBuilder implementation to use for processing the configuration; the configuration-specific
+     * CacheManagerBuilder implementation may emit a configuration-specific CacheManager implementation.
+     */
+    for (final ServiceCreationConfiguration<?> c : config.getServiceCreationConfigurations()) {
+      if (c instanceof CacheManagerConfiguration) {
+        cmBuilder = cmBuilder.with((CacheManagerConfiguration) c);
+      } else {
+        cmBuilder = cmBuilder.using(c);
+      }
+    }
+
+    for (final Map.Entry<String, CacheConfiguration<?, ?>> cc : config.getCacheConfigurations().entrySet()) {
+      cmBuilder = cmBuilder.withCache(cc.getKey(), cc.getValue());
+    }
+    cmBuilder = cmBuilder.withClassLoader(config.getClassLoader());
+
+    return specialization.cast(cmBuilder.build(false));
   }
 }
